+++ src/main/java/racing/controller/RacingController.java
@@ -3,12 +3,12 @@ package racing.controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racing.CarFactory;
+import racing.domain.CarFactory;
 import racing.controller.dto.request.RacingGameInfoRequest;
 import racing.controller.dto.response.RacingCarStateResponse;
 import racing.controller.dto.response.RacingGameResultResponse;
 import racing.domain.Cars;
-import racing.domain.service.RacingGameService;
+import racing.service.RacingGameService;
 
 import java.util.List;
 import java.util.stream.Collectors;
