+++ src/main/java/racing/repositroy/RacingGameDao.java
@@ -1,19 +1,17 @@
-package racing;
+package racing.repositroy;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
-import javax.annotation.PostConstruct;
 import java.sql.PreparedStatement;
 import java.sql.Statement;
 import java.sql.Timestamp;
 import java.util.Objects;
+import racing.repositroy.CarEntity;
 
 @Repository
 public class RacingGameDao {
